Don't share response objects r=amullings
In ResponseLists, if the schema is passed by reference, then it is
retrieved and used to construct the response object. If two
responses with different http status codes exist, they'll use the
same constructed response and overwrite fields. This change aims to
prevent that by copying the schema object instead.